redis: moving the firewall rule name into the validate package
