+++ azurerm/internal/services/redis/redis_firewall_rule_resource.go
@@ -3,7 +3,6 @@ package redis
 import (
 	"fmt"
 	"log"
-	"regexp"
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/services/redis/mgmt/2018-03-01/redis"
@@ -13,6 +12,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/tf"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/redis/validate"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/timeouts"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
 )
@@ -39,7 +39,7 @@ func resourceArmRedisFirewallRule() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: validateRedisFirewallRuleName,
+				ValidateFunc: validate.FirewallRuleName,
 			},
 
 			"redis_cache_name": {
@@ -178,14 +178,4 @@ func resourceArmRedisFirewallRuleDelete(d *schema.ResourceData, meta interface{}
 	}
 
 	return nil
-}
-
-func validateRedisFirewallRuleName(v interface{}, k string) (warnings []string, errors []error) {
-	value := v.(string)
-
-	if matched := regexp.MustCompile(`^\w+$`).Match([]byte(value)); !matched {
-		errors = append(errors, fmt.Errorf("%q may only contain alphanumeric characters and underscores", k))
-	}
-
-	return warnings, errors
 }
