+++ azurerm/internal/services/redis/validation_test.go
@@ -1,6 +1,8 @@
 package redis
 
-import "testing"
+import (
+	"testing"
+)
 
 func TestAccAzureRMRedisCacheFamily_validation(t *testing.T) {
 	cases := []struct {
@@ -93,46 +95,6 @@ func TestAccAzureRMRedisCacheBackupFrequency_validation(t *testing.T) {
 
 		if len(errors) != tc.ErrCount {
 			t.Fatalf("Expected the AzureRM Redis Cache Backup Frequency to trigger a validation error for '%d'", tc.Value)
-		}
-	}
-}
-
-func TestAzureRMRedisFirewallRuleName_validation(t *testing.T) {
-	cases := []struct {
-		Value    string
-		ErrCount int
-	}{
-		{
-			Value:    "ab",
-			ErrCount: 0,
-		},
-		{
-			Value:    "abc",
-			ErrCount: 0,
-		},
-		{
-			Value:    "webapp1",
-			ErrCount: 0,
-		},
-		{
-			Value:    "hello-world",
-			ErrCount: 1,
-		},
-		{
-			Value:    "hello_world",
-			ErrCount: 0,
-		},
-		{
-			Value:    "helloworld21!",
-			ErrCount: 1,
-		},
-	}
-
-	for _, tc := range cases {
-		_, errors := validateRedisFirewallRuleName(tc.Value, "azurerm_redis_firewall_rule")
-
-		if len(errors) != tc.ErrCount {
-			t.Fatalf("Expected the Redis Firewall Rule Name to trigger a validation error for '%s'", tc.Value)
 		}
 	}
 }
